refactor!(jlib-db): DBC API
- pageQuery params long -> int
  之所以从long改为int，是因为一般情况都查不到int最大值的页码和页大小
  而且使用long与Indexed的索引值大小不符，与FnRsRowMapping索引值大小也不符
  而且 DBC 中的execQuery能承担使用long分页的需求，
  其查询时若需要long大小的分页参数（当int装不下时，这种情况很罕见），则可以手动计算execQuery参数的分页sql然后查出来。
  总而言之，int大小的分页参数绝大部分情况够用了，而且能与各api统一
- 修改异常 ResultSetMappingFailException 为
  ResultSetCollectFailException，这更符合新的函数接口名

+++ jlib-db/src/main/java/io/github/baifangkual/jlib/db/DBC.java
@@ -217,7 +217,7 @@ public interface DBC {
      * @param <ROWS>   表示表中数据对象的类型
      * @return 表中符合分页要求的行
      */
-    <ROWS> ROWS tableData(String table, long pageNo, long pageSize,
+    <ROWS> ROWS tableData(String table, int pageNo, int pageSize,
                           FnResultSetCollector<? extends ROWS> fnRsMap);
 
     /**
