+++ jlib-db/src/main/java/io/github/baifangkual/jlib/db/trait/NoDBJustSchemaMetaProvider.java
@@ -28,7 +28,7 @@ public interface NoDBJustSchemaMetaProvider extends MetaProvider {
 
     <ROWS> ROWS tableData(Connection conn, String schema, String table,
                           Map<String, String> other,
-                          long pageNo, long pageSize,
+                          int pageNo, int pageSize,
                           FnResultSetCollector<? extends ROWS> fnResultSetCollector) throws Exception;
 
     private Tup2<String, Map<String, String>> unsafeGetSchemaAndOther(Cfg conf) {
@@ -59,7 +59,7 @@ public interface NoDBJustSchemaMetaProvider extends MetaProvider {
 
     @Override
     default <ROWS> ROWS tableData(Connection conn, Cfg config, String table,
-                                  long pageNo, long pageSize,
+                                  int pageNo, int pageSize,
                                   FnResultSetCollector<? extends ROWS> fnResultSetCollector) {
         try {
             Tup2<String, Map<String, String>> t2 = unsafeGetSchemaAndOther(config);
